+++ NEWS.md
@@ -1,5 +1,9 @@
-## gDRutils 1.5.13 - 2025-03-12
+## gDRutils 1.5.14 - 2025-03-12
 * improve logic in `get_assay_req_uniq_cols`
+
+## gDRutils 1.5.13 - 2025-03-07
+* add support for combination experiment in `cap_assay_infinities`
+* move `map_conc_to_standardized_conc` from gDRcore package
 
 ## gDRutils 1.5.12 - 2025-02-21
 * refactor `average_biological_replicates`